% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R, R/ui.R
\name{MPG Analysis}
\alias{MPG Analysis}
\title{MPG Analysis Server R code}
\arguments{
\item{slider1}{selects engine displacement size}

\item{slider2}{selects number of cylinders}

\item{showModel1}{determines if model will be displayed in graph}
}
\value{
a 'lm' object representing the linear model with the relevant values from mtcars

a 'lm' object representing the linear model with the relevant values from mtcars
}
\description{
This function runs a trend analysis of MPG against engine displacement and number of cylinders
the raw data is automatically filter and displayed as the displacement and/or number of cylinders
change.

This function displays the user interface of a trend analysis of MPG against engine displacement and number of cylinders
the raw data is automatically filter and displayed as the displacement and/or number of cylinders
change.
}
\details{
MPG Analysis

This function displays a tren analysis based MPG and engind displacement, filtered by
cylinder count and displacement size.

Using the mtcars dataframe to show trends of MPG based on the vehicles
engine displacement and cylinder size.

MPG Analysis

This function displays a tren analysis based MPG and engind displacement, filtered by
cylinder count and displacement size.

Using the mtcars dataframe to show trends of MPG based on the vehicles
engine displacement and cylinder size.
}
\seealso{
\code{lm, shiny, dplyr}

\code{lm}
}
\author{
Kiefer Maddex

Kiefer Maddex
}
